+++ CHANGELOG.md
@@ -7,3 +7,11 @@
 + CHANGELOG.md
 + optional blueprint field parameter `flip` for reverse order
 
+## 1.0.3
+
+### Added
++ optional blueprint field parameter `limit` for pagination
+
+### Changed
++ make use of the `pages/sidebar/subpages`-snippet to generate the list of subpages
+